requirements.txt
django
django-debug-toolbar
djangorestframework
django-filter
django-cors-headers
plaid-python
django-extensions
Werkzeug 